requirements.txt
Flask==2.0.2
tensorflow==2.13.0
Pillow==8.4.0
mysql-connector-python==8.0.28
